report viewer/pyproject.toml
[project]
name = "report-viewer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents~=1.3",
    "python-dotenv>=1.2.1",
]
